requirements.txt
google-generativeai
qdrant-client
neo4j
python-dotenv
tqdm
streamlit